+++ src/copoly.C
@@ -73,12 +73,12 @@ C_Oriented_Poly::~C_Oriented_Poly()
  ***************************************************/
 void C_Oriented_Poly::create_points(Port_3D &ref_port)
 {
-  int npoints = p_info->npoints;
+  unsigned int npoints = p_info->npoints;
   if (wpoints != NULL)
     delete [] wpoints;
   wpoints = new R_3DPoint[npoints];
   MYCHECK(wpoints != NULL);
-  for (int i =0;i<npoints;i++)
+  for (unsigned int i =0;i<npoints;i++)
     {
       R_3DPoint p = p_info->lpoints[i];
       p *= co_scaler;
@@ -175,7 +175,7 @@ void C_Oriented_Poly::set_poly_icolor(Vector &light)
       if (intensity > 1.0 )
 	intensity = 1.0;
       {
-	int col = base_color + (int) (intensity * ((float) color_range) );
+	int col = base_color + static_cast<int>(intensity * color_range);
 	/*
 	if (col <= base_color + color_range &&
 	    col >= base_color)
@@ -201,7 +201,7 @@ void C_Oriented_Poly::write(std::ostream &os)
       cc = '1';
     }
   os << '[' << ' ' << p_info->npoints  << ' ' << cc << ' ';
-  for (int i=p_info->npoints-1;i>=0;i--)
+  for (int i=static_cast<int>(p_info->npoints)-1;i>=0;i--)
     {
       if (!tflag)
 	{
@@ -228,7 +228,7 @@ std::ostream &operator <<(std::ostream &os, C_Oriented_Shape &cs)
 {
   os << '{' << '\n';
   os << cs.npolys << '\n';
-  for (int i=0;i<cs.npolys;i++)
+  for (unsigned int i=0;i<cs.npolys;i++)
     os << cs.o_polys[i] << '\n';
   os << '}' << '\n';
   return os;
@@ -251,7 +251,7 @@ int C_Oriented_Shape::create(C_ShapeInfo *ptr)
   o_polys = new C_Oriented_Poly[npolys];
   if (o_polys != NULL)
     {
-      for (int i=0;i<npolys;i++)
+      for (unsigned int i=0;i<npolys;i++)
 	{
 	  if (!o_polys[i].create(&info_ptr->polyinfos[i]))
 	    {
@@ -279,7 +279,7 @@ C_Oriented_Shape::~C_Oriented_Shape()
 
 std::istream &operator >>(std::istream &is, C_Oriented_Shape &cs)
 {
-  for (int i=0;i<cs.npolys;i++)
+  for (unsigned int i=0;i<cs.npolys;i++)
     is >> cs.o_polys[i];
   return is;
 }
@@ -287,7 +287,7 @@ std::istream &operator >>(std::istream &is, C_Oriented_Shape &cs)
 void C_Oriented_Shape::set_params(shape_params *prms)
 {
   flags = prms->flags;
-  for (int i=0;i<npolys;i++)
+  for (unsigned int i=0;i<npolys;i++)
     o_polys[i].set_params(&prms->p_params[i]);
 }
 
@@ -295,7 +295,7 @@ void C_Oriented_Shape::set_world_points(Port_3D &rf_port)
 {
   ref_port = &rf_port;
   maxlen = 0;
-  int i;
+  unsigned int i;
   visible = 1;
   for (i = 0; i <npolys;i++)
     {
@@ -320,7 +320,7 @@ void C_Oriented_Shape::set_world_points(Port_3D &rf_port)
 
 void C_Oriented_Shape::set_visible(Port_3D &theport)
 {
-  int i;
+  unsigned int i;
 
   if (frame_switch)
     {
@@ -347,7 +347,7 @@ void C_Oriented_Shape::set_visible(Port_3D &theport)
 
 void C_Oriented_Shape::set_poly_icolor(Vector &light)
 {
-  for (int i=0;i<npolys;i++)
+  for (unsigned int i=0;i<npolys;i++)
     {
       // do it "inline"
       C_Oriented_Poly &co = o_polys[i];
@@ -357,7 +357,7 @@ void C_Oriented_Shape::set_poly_icolor(Vector &light)
 	  if (intensity > 1.0 )
 	    intensity = 1.0;
 	  {
-	    int col = co.base_color + (int) (intensity * ((float) co.color_range - 1) );
+	    int col = co.base_color + static_cast<int>(intensity * (co.color_range - 1) );
 	    if (col <= co.base_color + co.color_range &&
 		col >= co.base_color)
 	      co.color = col;
@@ -371,18 +371,18 @@ void C_Oriented_Shape::set_poly_icolor(Vector &light)
 void C_Oriented_Shape::render_shadow(Port_3D &port, Vector &light)
 {
   MYCHECK(ref_port != NULL);
-  for (int i=0;i<npolys;i++)
+  for (unsigned int i=0;i<npolys;i++)
     {
       if (O_CheckPlaneEquation(light,o_polys[i].surface_normal,
 			       o_polys[i].plane_constant))
-	o_polys[i].render_shadow(port,(int) (flags & Z_CLIPME),0,
+	o_polys[i].render_shadow(port,static_cast<int>(flags & Z_CLIPME),0,
 				 light,0,*ref_port);
     }
 }
 
 void C_Oriented_Shape::draw(Port_3D &port, R_3DPoint &)
 {
-  int i;
+  unsigned int i;
   if (!visible)
     return;
   MYCHECK(ref_port != NULL);
@@ -400,13 +400,13 @@ void C_Oriented_Shape::draw(Port_3D &port, R_3DPoint &)
 
 void C_Oriented_Shape::create_points(Port_3D &ref_por)
 {
-  for (int i=0;i<npolys;i++)
+  for (unsigned int i=0;i<npolys;i++)
     o_polys[i].create_points(ref_por);
 }
 
 void C_Oriented_Shape::delete_points()
 {
-  for (int i=0;i<npolys;i++)
+  for (unsigned int i=0;i<npolys;i++)
     {
       if (o_polys[i].wpoints != NULL)
 	delete [] o_polys[i].wpoints;
@@ -416,6 +416,6 @@ void C_Oriented_Shape::delete_points()
 
 void C_Oriented_Shape::set_poly_color(int n, int color)
 {
-  if (n >= 0 && n < npolys)
+  if (n >= 0 && static_cast<unsigned int>(n) < npolys)
     o_polys[n].color = color;
 }
